+++ lib/percona_migrator/railtie.rb
@@ -25,8 +25,19 @@ module PerconaMigrator
           # @param direction [Symbol] :up or :down
           def migrate(direction)
             reconnect_with_percona
+            include_foreigner if defined?(Foreigner)
+
             ::Lhm.migration = self
             original_migrate(direction)
+          end
+
+          # Includes the Foreigner's Mysql2Adapter implemention in
+          # PerconaMigratorAdapter to support foreign keys
+          def include_foreigner
+            Foreigner::Adapter.safe_include(
+              :PerconaMigratorAdapter,
+              Foreigner::ConnectionAdapters::Mysql2Adapter
+            )
           end
 
           # Make all connections in the connection pool to use PerconaAdapter
@@ -38,8 +49,8 @@ module PerconaMigrator
             )
           end
 
-          # It executes the passed statement through the PerconaAdapter if it's
-          # an alter statement. It uses the mysql adapter otherwise.
+          # It executes the passed statement through the PerconaMigratorAdapter
+          # if it's an alter statement. It uses the mysql adapter otherwise.
           #
           # This is because +pt-online-schema-change+ is intended for alter
           # statements only.